chore: Move tests with source files
