+++ packages/feathers/fixtures/index.ts
@@ -2,8 +2,8 @@ import { createServerAdapter } from '@whatwg-node/server'
 import { createServer } from 'node:http'
 import { TestService } from './fixture.js'
 
-import { feathers, Application, Params } from '../../src/index.js'
-import { createHandler } from '../../src/http/index.js'
+import { feathers, Application, Params } from '../src/index.js'
+import { createHandler } from '../src/http/index.js'
 
 export * from './client.js'
 export * from './rest.js'
