+++ packages/feathers/src/declarations.test.ts
