+++ packages/feathers/src/errors.test.ts
