+++ packages/feathers/src/hooks/compose.test.ts
@@ -1,7 +1,7 @@
 // Adapted from koa-compose (https://github.com/koajs/compose)
 import { describe, expect, it } from 'vitest'
 import assert from 'assert'
-import { compose, NextFunction } from '../../src/hooks/index.js'
+import { compose, NextFunction } from './index.js'
 
 function wait(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms || 1))
