+++ packages/feathers/src/hooks/decorator.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import assert from 'assert'
-import { HookContext, hooks, middleware, NextFunction } from '../../src/hooks/index.js'
+import { HookContext, hooks, middleware, NextFunction } from './index.js'
 
 describe('feathers/hooks decorator', () => {
   it('hook decorator on method and classes with inheritance', async () => {
