+++ packages/feathers/src/hooks/feathers/around.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'vitest'
 import assert from 'assert'
-import { feathers, Params, ServiceInterface } from '../../../src/index.js'
+import { feathers, Params, ServiceInterface } from '../../index.js'
 
 describe('`around` hooks', () => {
   it('around hooks can set hook.result which will skip service method', async () => {
